tests/wpt/web-platform-tests/workers/semantics/navigation/001.html

<!--

/*
-->
<!doctype html>
<title>navigating</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
var date;
var newDate;
</script>
<iframe></iframe>
<script>
var iframe = document.querySelector('iframe');
onload = function() {
  iframe.src = "001-1.html?" + Math.random();
};
var start_test = function() {
  window[0].document.links[0].click();
};
var after_load = function() {
  history.back();
  newDate = new Date();
  setTimeout(function() {
    assert_greater_than(Number(date), Number(newDate));
    assert_equals(window[0].document.title, '001-1');
    done();
  }, 2000);
};
</script>
<!--
*/
//-->
